+++ packages/riverpod/lib/src/framework.dart
@@ -12,6 +12,7 @@ part 'framework/always_alive.dart';
 part 'framework/auto_dispose.dart';
 part 'framework/async_selector.dart';
 part 'framework/provider_base.dart';
+part 'framework/element.dart';
 part 'framework/container.dart';
 part 'framework/family.dart';
 part 'framework/listen.dart';
